+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No changes yet.
+### Fixed
+- Observability middleware will always emit an error code if the returned error
+  is from the `yarpcerrors` package.
 
 ## [1.37.4] - 2019-05-02
 ### Fixed